Home page upgraded L to D
