+++ src/App.js
@@ -3,25 +3,31 @@ import { createGlobalStyle, ThemeProvider } from 'styled-components';
 import Navigation from './Navigation';
 
 const darkTheme = {
+  fontcolor: "#fff",
   headerBackground: "#2E2E2E",
   bodyBackground: "#393939",
-  fontcolor: "#fff",
-  loginHeader: "FEFEFE",
+  loginContainerBackground: "#474747",
+  loginHeader: "#FEFEFE",
   buttonTextActive: "#fff",
   buttonTextInActive: "#D0D0D0",
   buttonBackgroundPrimary: "linear-gradient(225deg, #F16648, #C1005B)",
   buttonBackgroundSecondary: "linear-gradient(225deg, #9C2441, #53506D)",
+  textBoxBackground: "#616161",
 };
 
 
 const lightTheme = {
-  fontcolor: "black",
+  fontcolor: "#050505",
   loginHeader:"transparent",
+  loginContainerBackground: "#fff",
   buttonTextActive: "#fff",
   headerBackground:"linear-gradient(90deg, #FFC1C8 2.88%, #FF9FA9 51.74%, #FFBFC6 100%)",
   bodyBackground: "#fff",
   buttonTextActive: "#fff",
   buttonTextInActive: "#D0D0D0",
+  buttonBackgroundPrimary: "linear-gradient(225deg, #F16648, #C1005B)",
+  buttonBackgroundSecondary: "linear-gradient(225deg, #9C2441, #53506D)",
+  textBoxBackground: "#E9E9E9",
 
 };
 
@@ -51,7 +57,7 @@ body {
 `;
 
   return (
-    <ThemeProvider theme={darkTheme}>
+    <ThemeProvider theme={lightTheme}>
       <GlobalStyles />
       <Navigation />
     </ThemeProvider>
