+++ src/components/TextField.jsx
@@ -1,22 +1,22 @@
 import React from 'react';
 
 function TextField(props) {
+
+
   const textStyle = {
       width: "100%",
       height: "100%",
       backgroundColor: "transparent",
-      fontSize: "15px",
       fontSize: "14px",
-      fontSize: "15px",
-      fontSize: "14px",
-      color: "#fff",
       outline: "none",
       border: "none",
       paddingLeft: "10px",
-      
-  }
+      color: "black",
+  };
+
+
   return(
-     <input type={props.type} placeholder = {props.placeholder} style={textStyle} color={"#fff"} />
+     <input type={props.type} placeholder = {props.placeholder} style={textStyle} />
   );
 }
