+++ src/index.css
@@ -13,16 +13,6 @@ code {
    
 }
 
-
-::placeholder{
-  color: #999999;
-}
-::placeholder{
-  font-size: 14px;
-  color:white;
-}
-
-
 ::-webkit-scrollbar {
   width: 7px;
 }
