+++ src/styles/HeadingFonts.js
@@ -1,7 +1,7 @@
 import styled from 'styled-components';
 
 export const LoginSignupHeader = styled.p`
-   color: #fff;
+   color: ${ (props) => props.theme.fontcolor };
    font-size: 30px;
    font-weight: 400;
