+++ src/styles/LoginHeaderStyle.js
@@ -3,6 +3,7 @@ import styled from 'styled-components';
 export const HeaderContainer = styled.div`
   width: 100%;
   height: 9vh;
+  background-image: ${(props) => props.theme.headerBackground}; 
   background-color: ${(props) => props.theme.headerBackground}; 
   min-height: 9vh;
   align-items: center;
